add some change

+++ controllers/users/user-wallet.controller.js
@@ -69,37 +69,44 @@ export const buyCoin = async (req, res) => {
     const id = req.params.id;
     const users = req.user_id;
     const order = await Order.findById({ _id: id });
-    const taxs = await Tax.findOne({});
-    const fromUserWallet = await UserWallet.findOne({ users: order.users });
-    const userWallet = await UserWallet.findOne({ users });
-    if (userWallet.price >= order.price) {
-      userWallet.coins += parseInt(order.coin);
-      userWallet.price -= parseInt(order.price);
-      await userWallet.save();
-
-      fromUserWallet.price += (parseInt(order.price) * taxs.taxs) / 100;
-      await fromUserWallet.save();
-
-      const transaction = new Transaction({
-        users: users,
-        from: order.users,
-        transaction: "Payment Received",
-        price: parseInt(order.price),
-        received: true,
-      });
-      const transactions = await transaction.save();
-      if (transactions) {
-        await Order.findByIdAndDelete({ _id: id });
-        return res.send({
-          success: true,
-          message: "Coin Buy Successfully",
-        });
-      }
-    } else {
+    if (order.users == users) {
       return res.send({
         success: false,
-        message: "Don't have enough money in wallet to buy this order",
+        message: "You cannot buy your order",
       });
+    } else {
+      const taxs = await Tax.findOne({});
+      const fromUserWallet = await UserWallet.findOne({ users: order.users });
+      const userWallet = await UserWallet.findOne({ users });
+      if (userWallet.price >= order.price) {
+        userWallet.coins += parseInt(order.coin);
+        userWallet.price -= parseInt(order.price);
+        await userWallet.save();
+
+        fromUserWallet.price += (parseInt(order.price) * taxs.taxs) / 100;
+        await fromUserWallet.save();
+
+        const transaction = new Transaction({
+          users: users,
+          from: order.users,
+          transaction: "Payment Received",
+          price: parseInt(order.price),
+          received: true,
+        });
+        const transactions = await transaction.save();
+        if (transactions) {
+          await Order.findByIdAndDelete({ _id: id });
+          return res.send({
+            success: true,
+            message: "Coin Buy Successfully",
+          });
+        }
+      } else {
+        return res.send({
+          success: false,
+          message: "Don't have enough money in wallet to buy this order",
+        });
+      }
     }
   } catch (error) {
     return res.send({
